代码告警清理，不要混用有符号数和无符号数

+++ window_scene/session_manager/src/scene_session_manager.cpp
@@ -414,7 +414,7 @@ void SceneSessionManager::ConfigFreeMultiWindow()
         }
         item = freeMultiWindowConfig["maxMainFloatingWindowNumber"];
         if (GetSingleIntItem(item, param) && (param > 0)) {
-            systemConfig_.freeMultiWindowConfig_.maxMainFloatingWindowNumber_ = param;
+            systemConfig_.freeMultiWindowConfig_.maxMainFloatingWindowNumber_ = static_cast<uint32_t>(param);
         }
     }
 }
